# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Luxe Clothes Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc]
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Coupons (public validate) ─────────────────────────────────────
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin-products]
      summary: Admin list products (all, with pagination)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated admin products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductList"
    post:
      operationId: adminCreateProduct
      tags: [admin-products]
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products/{id}:
    get:
      operationId: adminGetProduct
      tags: [admin-products]
      summary: Get a single product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: adminUpdateProduct
      tags: [admin-products]
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteProduct
      tags: [admin-products]
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Categories ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Settings ─────────────────────────────────────────────────────
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        stock:
          type: integer
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, price, stock, sizes, colors, images, featured, createdAt]
    ProductList:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, limit, totalPages]
    AdminProductList:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, limit, totalPages]
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        price:
          type: number
          minimum: 0
        salePrice:
          type: ["number", "null"]
        stock:
          type: integer
          minimum: 0
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
          default: false
      required: [name, price, stock, sizes, colors, images]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        stock:
          type: integer
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean

    # ── Coupon ────────────────────────────────────────────────────────
...